requirements.txt
ccxt==4.4.26
pandas==2.2.3
numpy==2.2.3
ta==0.11.0
pyyaml==6.0.2
python-dotenv==1.0.1
rich==13.9.4
websockets==14.2
pytest==8.3.4
